datatype changes and mileage pom file

+++ java/commoncomponent/common.serialization.deserialization/src/main/java/net/atos/daf/ct2/pojo/standard/SparseMatrix.java
@@ -15,9 +15,9 @@ public class SparseMatrix  implements Serializable {
   @JsonProperty(value = "ord")
   private Long ord;
   @JsonProperty(value = "A")
-  private Integer[] a;
+  private Long[] a;
   @JsonProperty(value = "IA")
-  private Integer[] ia;
+  private Long[] ia;
   @JsonProperty(value = "JA")
-  private Integer[] ja;
+  private Long[] ja;
 }
